Removes app/pictures directory.
This introduces the dev_pictures directory, which will allow to be run with the development pics, used in tests.

The goal is to move towards providing a directory at build time to be copied into the dist directory, to copy how it works on live servers. The old pictures/pictures.json is being replaced by a relative manifest.json within the pictures directory.